addBooksFunction

+++ Fptbook/Models/Entity/Book.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
         public int Quanlity { get; set; }
         public string Description {get; set; }
         public DateTime DateCreated { get; set; }
@@ -16,7 +16,7 @@
         public string ImagePath { get; set; }
         public int CategoryId { get; set; }
         public Category Category { get; set; }
-        public int Pages { get; set; }
+        public double Pages { get; set; }
 
     }
 }
